polymarket-btc-bot/requirements.txt
# Polymarket
py-clob-client==0.34.6

# Blockchain
web3>=6.0.0

# WebSocket / HTTP
websockets>=12.0
aiohttp>=3.9.0

# Database
aiosqlite>=0.20.0

# Telegram
python-telegram-bot>=21.0

# Config / validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Logging
structlog>=24.0.0

# Utilities
python-dateutil>=2.8.2
